remove_old_signer
