+++ Cargo.toml
@@ -2,7 +2,6 @@
 members = [
     "src/icrc-client",
     "src/did",
-    "src/eth-signer-alloy",
     "src/eth-signer",
     "src/eth-signer/tests/test_canister",
     "src/ethereum-json-rpc-client",
