--- src/eth-signer-alloy/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-categories = ["cryptography::cryptocurrencies"]
-description = "API types definition for EVM canister"
-include = ["src/**/*", "../../LICENSE", "../../README.md"]
-name = "eth-signer-alloy"
-
-authors.workspace = true
-homepage.workspace = true
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-
-[features]
-ic_sign = ["ic-canister", "ic-exports"]
-
-[dependencies]
-alloy = { workspace = true, features = ["network", "signer-local"]}
-async-trait = { workspace = true }
-candid = { workspace = true }
-did = { path = "../did" }
-hex = { workspace = true }
-ic-stable-structures = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-sha2 = { workspace = true }
-thiserror = { workspace = true }
-
-# Dependencies for ic-siginig
-ic-exports = { workspace = true, optional = true }
-ic-canister = { workspace = true, optional = true }
-
-[dev-dependencies]
-ic-exports = { workspace = true, features = ["pocket-ic-tests"]}
-rand = { workspace = true }
-tokio = { workspace = true }
