--- src/eth-signer-alloy/README.md
@@ -1,1 +1,0 @@
-# Eth Signer
