--- src/eth-signer-alloy/tests/it.rs
@@ -1,1 +1,0 @@
-mod pocket_ic_tests;
