--- src/eth-signer-alloy/tests/pocket_ic_tests/mod.rs
@@ -1,2 +1,0 @@
-mod ic_sign_test_canister;
-mod wasm_utils;
