--- src/eth-signer-alloy/tests/test_canister/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "ic-sign-test-canister"
-
-authors.workspace = true
-homepage.workspace = true
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-
-[features]
-default = []
-export-api = []
-
-
-[dependencies]
-candid = { workspace = true }
-eth-signer = { path = "../../", features = ["ic_sign"] }
-ic-exports = { workspace = true }
-ic-canister = { workspace = true }
-ethers-core = { workspace = true }
