--- src/eth-signer-alloy/tests/test_canister/README.md
@@ -1,10 +1,0 @@
-# Canister to test management canister ECDSA signing.
-
-## Test
-The `sign_and_check` test creates a transaction, signs it using management canister and recovers `from` address from signature.
-
-## Running
-To perform the test, use `/scripts/dfx_test.sh` script.
-
-**NOTE: `dfx` should be started**
-
