+++ src/eth-signer/Cargo.toml
@@ -15,15 +15,14 @@ repository.workspace = true
 ic_sign = ["ic-canister", "ic-exports"]
 
 [dependencies]
-alloy = { workspace = true }
+alloy = { workspace = true, features = ["network", "signer-local"]}
 async-trait = { workspace = true }
 candid = { workspace = true }
 did = { path = "../did" }
-ethereum-types = { workspace = true }
-ethers-core = { workspace = true }
 hex = { workspace = true }
 ic-stable-structures = { workspace = true }
 serde = { workspace = true }
+serde_json = { workspace = true }
 sha2 = { workspace = true }
 thiserror = { workspace = true }
